requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
elastic-transport==9.1.0
elasticsearch==9.1.1
Flask==3.1.2
flask-cors==6.0.1
geopandas==1.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pyogrio==0.11.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
shapely==2.1.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
